+++ src/main/java/dinkplugin/notifiers/MetaNotifier.java
@@ -1,5 +1,7 @@
 package dinkplugin.notifiers;
 
+import com.google.gson.Gson;
+import com.google.gson.JsonSyntaxException;
 import dinkplugin.domain.AchievementDiary;
 import dinkplugin.message.NotificationBody;
 import dinkplugin.message.NotificationType;
@@ -7,7 +9,9 @@ import dinkplugin.message.templating.Replacements;
 import dinkplugin.message.templating.Template;
 import dinkplugin.notifiers.data.LoginNotificationData;
 import dinkplugin.notifiers.data.Progress;
+import dinkplugin.util.SerializedPet;
 import dinkplugin.util.Utils;
+import lombok.extern.slf4j.Slf4j;
 import net.runelite.api.Experience;
 import net.runelite.api.GameState;
 import net.runelite.api.Skill;
@@ -15,18 +19,26 @@ import net.runelite.api.VarPlayer;
 import net.runelite.api.Varbits;
 import net.runelite.api.events.GameStateChanged;
 import net.runelite.client.callback.ClientThread;
+import net.runelite.client.config.ConfigManager;
+import net.runelite.client.config.RuneLiteConfig;
+import net.runelite.client.plugins.chatcommands.ChatCommandsPlugin;
 import org.apache.commons.lang3.StringUtils;
 import org.jetbrains.annotations.VisibleForTesting;
 
 import javax.inject.Inject;
 import javax.inject.Singleton;
+import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.atomic.AtomicReference;
 
+@Slf4j
 @Singleton
 public class MetaNotifier extends BaseNotifier {
+    static final @VisibleForTesting String RL_CHAT_CMD_PLUGIN_NAME = ChatCommandsPlugin.class.getSimpleName().toLowerCase();
     static final @VisibleForTesting int INIT_TICKS = 10; // 6 seconds after login
 
     private final AtomicInteger loginTicks = new AtomicInteger(-1);
@@ -34,6 +46,12 @@ public class MetaNotifier extends BaseNotifier {
 
     @Inject
     private ClientThread clientThread;
+
+    @Inject
+    private ConfigManager configManager;
+
+    @Inject
+    private Gson gson;
 
     @Override
     public boolean isEnabled() {
@@ -123,7 +141,8 @@ public class MetaNotifier extends BaseNotifier {
             new LoginNotificationData.SkillData(experienceTotal, levelTotal, skillLevels, skillExperience),
             Progress.of(questsCompleted, questsTotal),
             Progress.of(questPoints, questPointsTotal),
-            new LoginNotificationData.SlayerData(slayerPoints, slayerStreak)
+            new LoginNotificationData.SlayerData(slayerPoints, slayerStreak),
+            getPets()
         );
         createMessage(false, NotificationBody.builder()
             .type(NotificationType.LOGIN)
@@ -132,6 +151,30 @@ public class MetaNotifier extends BaseNotifier {
             .playerName(playerName)
             .build()
         );
+    }
+
+    @VisibleForTesting
+    List<SerializedPet> getPets() {
+        if ("false".equals(configManager.getConfiguration(RuneLiteConfig.GROUP_NAME, RL_CHAT_CMD_PLUGIN_NAME)))
+            return null;
+
+        String json = configManager.getRSProfileConfiguration("chatcommands", "pets2");
+        if (json == null || json.isEmpty())
+            return null;
+
+        int[] petItemIds;
+        try {
+            petItemIds = gson.fromJson(json, int[].class);
+        } catch (JsonSyntaxException e) {
+            log.info("Failed to deserialize owned pet IDs", e);
+            return null;
+        }
+
+        List<SerializedPet> pets = new ArrayList<>(petItemIds.length);
+        for (int itemId : petItemIds) {
+            pets.add(new SerializedPet(itemId, client.getItemDefinition(itemId).getMembersName()));
+        }
+        return pets;
     }
 
 }